# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nexus Team Task Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity across all projects
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
  /projects/{id}/stats:
    get:
      operationId: getProjectStats
      tags: [dashboard]
      summary: Get task stats for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

  # ── User profile sync ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [task_created, task_updated, task_completed, task_assigned, comment_added, member_added, project_created]
        projectId:
          type: integer
        projectName:
          type: string
        taskId:
          type: ["integer", "null"]
        taskTitle:
          type: ["string", "null"]
        actorClerkUserId:
          type: string
        actorName:
          type: string
        actorAvatarUrl:
          type: ["string", "null"]
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - projectId
        - projectName
        - taskId
        - taskTitle
        - actorClerkUserId
        - actorName
        - actorAvatarUrl
        - description
        - createdAt
    ProjectStats:
      type: object
      properties:
        projectId:
          type: integer
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        inProgressTasks:
          type: integer
        overdueTasks:
          type: integer
        completionRate:
          type: number
        tasksByStatus:
          type: object
          properties:
            todo:
              type: integer
            in_progress:
              type: integer
            in_review:
              type: integer
            done:
              type: integer
          required:
            - todo
            - in_progress
            - in_review
            - done
        tasksByPriority:
          type: object
          properties:
            low:
              type: integer
            medium:
              type: integer
            high:
              type: integer
            urgent:
              type: integer
          required:
            - low
            - medium
            - high
            - urgent
      required:
        - projectId
        - totalTasks
        - completedTasks
        - inProgressTasks
        - overdueTasks
        - completionRate
        - tasksByStatus
        - tasksByPriority